Changed int to long because they are the same size even on a 64-bit Windows platform.
git-svn-id: https://svn.code.sf.net/p/jazzplusplus/code/trunk/jazz@907 d001d500-7bcc-4352-ac0f-06d6f3fd1383

+++ src/Audio.cpp
@@ -622,7 +622,7 @@ void JZSampleSet::AdjustAudioLength(JZTrack* pTrack, long TicksPerMinute)
 
 //-----------------------------------------------------------------------------
 //-----------------------------------------------------------------------------
-void JZSampleSet::StartPlay(long clock)
+void JZSampleSet::StartPlay(int clock)
 {
   ReloadSamples();
